move add and delete card to level of Board
